+++ app/blog/[slug]/page.tsx
@@ -2,7 +2,7 @@
 import { fetchBlogData } from "@src/services";
 import type { Metadata, ResolvingMetadata } from "next";
 import Blog from "@components/Blog/Blog";
-import { BackButton } from "@components/buttons";
+import { BackButton } from "@components/ui/button";
 type Props = {
     params: Promise<{ slug: string }>;
     searchParams: Promise<Record<string, string | string[] | undefined>>;
