+++ src/components/Blog/manage/authors/AuthorForm.tsx
@@ -1,8 +1,7 @@
 // src/components/Blog/manage/authors/AuthorForm.tsx
 "use client";
 import React, { forwardRef, type ChangeEvent } from "react";
-import EditableField from "@components/forms/EditableField";
-import EditableTextArea from "@components/forms/EditableTextArea";
+import { EditableField, EditableTextArea } from "@components/ui/form";
 import EntityFormShell from "@components/Blog/manage/EntityFormShell";
 import { type AuthorFormType, initialAuthorForm, useAuthorForm } from "@entities/models/author";
