+++ src/components/Blog/manage/components/FormActionButtons.tsx
@@ -1,6 +1,5 @@
 import ActionButtons from "./buttons/ActionButtons";
-import { EditButton, DeleteButton } from "@components/buttons";
-import ButtonBase from "@components/buttons/ButtonBase";
+import { EditButton, DeleteButton, ButtonBase } from "@components/ui/button";
 type IdLike = string | number;
 
 interface FormActionButtonsProps {
