+++ src/components/Profile/UserNameManager.tsx
@@ -3,7 +3,7 @@
 import React, { useEffect, useState, useCallback } from "react";
 import { Authenticator, useAuthenticator } from "@aws-amplify/ui-react";
 import "@aws-amplify/ui-react/styles.css";
-import EntityEditor from "@components/forms/EntityEditor";
+import { EntityEditor } from "@components/ui/form";
 import { label as fieldLabel } from "./utilsUserName";
 import PersonIcon from "@mui/icons-material/Person";
 import { useUserNameForm } from "@entities/models/userName/hooks";
