--- src/components/buttons/index.ts
@@ -1,2 +1,0 @@
-export * from "./Buttons";
-export * from "./buttonStyles";
