+++ src/components/ui/button/index.ts
@@ -1,0 +1,4 @@
+export * from "./Buttons";
+export * from "./buttonStyles";
+export { default as ButtonBase } from "./ButtonBase";
+export type { ButtonBaseProps } from "./ButtonBase";
